[GmlSt] Move scalarization pass to gml_st.
There is no good reason for scalarization pass not to be in the same directory with vectorization. Also, if it's not there, we are getting a circular dependence when adding scalarization pass to gml_st_cpu_pipeline.

Also remove gml_st_pipeline.cc, which was removed already, but reappeared after a bad merge during llvm integrate.

PiperOrigin-RevId: 497025917